requirements.txt
python-dotenv
openai
pyupbit
